Move ScrollToTop to layout components
add tabIndex functionality
