+++ src/app/(projects)/_projectlayout_/index.ts
@@ -1,0 +1,5 @@
+import Header from './Header';
+import ProjectCategoryNavItem from './ProjectCategoryNavItem';
+import ScrollToTop from './ScrollToTop';
+
+export { Header, ProjectCategoryNavItem, ScrollToTop };
